requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
datasets==2.20.0
dill==0.3.8
evaluate==0.4.2
fastembed-gpu==0.3.3
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
inquirerpy==0.3.4
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnxruntime-gpu==1.18.1
packaging==24.1
pandas==2.2.2
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge_score==0.1.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
snowballstemmer==2.2.0
sympy==1.13.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
